requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
joblib
requests
beautifulsoup4
python-telegram-bot
plotly
python-dotenv
loguru
apscheduler
